Printing code has been moved to seperate class - UserInterface. Reading from the database in CSVDatabase has been properly refactored.
