+++ Chirp.CLI.csproj
@@ -9,8 +9,6 @@
 
   <ItemGroup>
     <PackageReference Include="CsvHelper" Version="33.0.1" />
-    <PackageReference Include="Chirp.CLI.SimpleDB" Version="5.6.1" />
-    <PackageReference Include="Chirp.CLI.Client" Version="5.6.1" />
   </ItemGroup>
 
 </Project>
