+++ Chirp.CLI.generated.sln
